Test compiling libsrc using gcc
```
$ make -C libsrc
$ gcc -nostdlib -Iinclude foo.c lib/crt0.a lib/libc.a
```

+++ libsrc/stdio/vsprintf.c
@@ -1,5 +1,5 @@
 #include "stdio.h"
 
 int vsprintf(char *buf, const char *fmt, va_list ap) {
-  return vsnprintf(buf, -1UL, fmt, ap);
+  return vsnprintf(buf, (size_t)-1, fmt, ap);
 }
